Add partial implementation of controlling settings with FS
This fix passes foot switch presses on to the gui task and
also to the User Settings. You can now advance through menu
settings with a single press.

Still left to be done is supporting double-press to send a
press event to a button.

Long press when in tuning mode to activate user settings.